Implement Search logic - Get value from autocomplete - Create single asset page - Redirect user to single asset page
