+++ src/components/search-bar/index.tsx
@@ -1,19 +1,23 @@
-import React from "react";
+import React, { useState } from "react";
 import { Stack, Autocomplete, TextField } from "@mui/material";
-import { useStyles } from "./styles";
 import { ISingleAsset } from "../../common/types/assets";
 import { useAppSelector } from "../../utils/hook";
+import { useNavigate } from "react-router-dom";
 
 const SearchBarComponent = () => {
+  const [selectedItem, setSelectedItem] = useState<string | null>("");
+  const navigate = useNavigate();
   const assetsArray: ISingleAsset[] = useAppSelector(
     (state) => state.assets.assets,
   );
-
-  const classes = useStyles();
   return (
     <Stack sx={{ width: 300 }}>
       <Autocomplete
-        freeSolo
+        value={selectedItem}
+        onChange={(e: any, value: string | null) => {
+          navigate(`single/${value}`);
+          setSelectedItem(null);
+        }}
         renderInput={(element) => (
           <TextField
             {...element}
