+++ src/pages/settings/index.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 
-const SettingsComponent = () => {
-    return (
-        <div>
-            <h1>Settings page</h1>
-        </div>
-    );
+const SettingsPage = () => {
+  return (
+    <div>
+      <h1>Settings page</h1>
+    </div>
+  );
 };
 
-export default SettingsComponent;
+export default SettingsPage;
