+++ src/pages/watchlist/index.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 
-const WatchlistComponent = () => {
-    return (
-        <div>
-            <h1>Watchlist</h1>
-        </div>
-    );
+const WatchlistPage = () => {
+  return (
+    <div>
+      <h1>Watchlist</h1>
+    </div>
+  );
 };
 
-export default WatchlistComponent;
+export default WatchlistPage;
